+++ apps/web/src/components/Menu/index.tsx
@@ -3,7 +3,7 @@ import { useRouter } from 'next/router'
 import { Menu as UikitMenu, NextLinkFromReactRouter, footerLinks } from '@pancakeswap/uikit'
 import { useTranslation, languageList } from '@pancakeswap/localization'
 import PhishingWarningBanner from 'components/PhishingWarningBanner'
-// import { NetworkSwitcher } from 'components/NetworkSwitcher'
+import { NetworkSwitcher } from 'components/NetworkSwitcher'
 import useTheme from 'hooks/useTheme'
 // import { useCakeBusdPrice } from 'hooks/useBUSDPrice'
 import { usePhishingBannerManager } from 'state/user/hooks'
@@ -42,7 +42,7 @@ const Menu = (props) => {
         rightSide={
           <>
             <GlobalSettings mode={SettingsMode.GLOBAL} />
-            {/* <NetworkSwitcher /> */}
+             <NetworkSwitcher />
             <UserMenu />
           </>
         }
