+++ apps/web/src/components/NetworkModal/PageNetworkSupportModal.tsx
@@ -23,7 +23,7 @@ export function PageNetworkSupportModal() {
   const foundChain = useMemo(() => chains.find((c) => c.id === chainId), [chainId])
   const historyManager = useHistory()
 
-  const lastValidPath = historyManager?.history?.find((h) => ['/swap', 'liquidity', '/', '/info'].includes(h))
+  const lastValidPath = historyManager?.history?.find((h) => ['/swap', 'liquidity', '/', '/info', '/claim'].includes(h))
 
   const menuItems = useMenuItems()
   const { pathname, push } = useRouter()
