+++ apps/web/src/pages/claim.tsx
@@ -1,3 +1,11 @@
+import { CHAIN_IDS } from 'utils/wagmi'
 import Claim from '../views/Claim'
 
+
+const ClaimPage = () => {
+  return  <Claim />
+}
 export default Claim
+
+
+ClaimPage.chains = CHAIN_IDS
